+++ src/WorkItemSearchAndLock/src/app/app.component.css
@@ -10,6 +10,10 @@
     margin: 0; 
 }
 
+.center-align{
+    text-align: center;
+}
+
 .btn {
     text-align: center;
     padding: 7px;
@@ -17,15 +21,15 @@
     cursor: pointer;
 }
 
-.substepList {
+.subStepList {
     list-style-type: none;
 }
 
-.substepTable {
+.subStepTable {
     clear: both;
 }
 
-.substepTable td, th {
+.subStepTable td, th {
     padding-right: 20px;
 }
 
@@ -38,6 +42,6 @@
     margin-right: 20px;
 }
 
-.statistics .stro {
+.statistics .strong {
     font-weight: bold;
 }
